+++ CHANGELOG.md
@@ -50,7 +50,7 @@
 
 ### General
 
-- [ ] There is now an enforced level cap.
+- [x] There is now an enforced level cap.
 - [x] Battle style is now always Set.
 - [ ] All trainer battles now scale with the number of badges the player has.
 - [ ] No items in battle for the player.